fix: bug fixes, UX enhancements, and comprehensive E2E tests
## Backend Fixes
- Add 30-second query timeout protection (database.py)
- Fix ORM model type mismatch: lncrna_start/end Integer type (models.py)
- Add security documentation for SQL query whitelisting

## Frontend Enhancements
- ErrorState: Add HTTP status-specific icons, dev mode stack trace
- LoadingState: Add customizable message/tip props
- Export: Add popup blocker detection, success/error feedback
- Fix unused import in LncRNAChIPSeqOverlapTable

## E2E Tests (162 new tests)
- chipseq-overlap-full-flow.spec.ts: Complete filter→browse→export flow
- regulations-comprehensive.spec.ts: Full regulations page coverage
- genome-browser.spec.ts: IGV.js integration tests
- network-visualization.spec.ts: Cytoscape network tests
- error-states-global.spec.ts: Global error handling tests

## Documentation
- Add BUG_FIX_REPORT_2025_12_08.md
- Add E2E test reports and documentation

+++ frontend/backend/app/models/models.py
@@ -151,8 +151,8 @@ class Regulation(Base):
     best_peak_start = Column(BigInteger)
     best_peak_end = Column(BigInteger)
     best_site_ba = Column(Numeric(10, 4))
-    lncrna_start = Column(BigInteger)  # 与 dna_start/end 保持一致，避免大坐标溢出
-    lncrna_end = Column(BigInteger)    # 与 dna_start/end 保持一致，避免大坐标溢出
+    lncrna_start = Column(Integer)  # Matches database schema (integer type)
+    lncrna_end = Column(Integer)    # Matches database schema (integer type)
     dna_start = Column(BigInteger)
     dna_end = Column(BigInteger)
     binding_affinity = Column(Numeric(10, 4))
